requirements.txt
absl-py==0.8.1
astor==0.8.0
atari-py==0.2.6
baselines==0.1.6
Click==7.0
cloudpickle==1.2.2
cycler==0.10.0
future==0.18.0
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.1
gym==0.15.3
gym-retro==0.7.0
h5py==2.10.0
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mpi4py==3.0.2
numpy==1.17.2
opencv-python==4.1.1.26
pandas==0.25.1
Pillow==6.2.0
pkg-resources==0.0.0
protobuf==3.10.0
pyglet==1.3.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
scipy==1.3.1
six==1.12.0
stable-baselines==2.8.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.36.1
Werkzeug==0.16.0
wrapt==1.11.2
